requirements.txt
Flask
Flask-SQLAlchemy
Flask-Mail
python-dotenv
openai
Pillow
PyMuPDF
python-docx
pandas
openpyxl
